Renamed component files & folders to suit naming conventions
