+++ src/client/components/stage-details/stage-details.story.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import StageDetails from './StageDetails.component';
+import StageDetails from './stage-details.component';
 import { storiesOf } from '@storybook/react';
 
 storiesOf('Accordion Options', module).add('Stage Details', () => {
