+++ src/client/components/stage-details/stage-details.style.css
